+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Features
 * Initial support for Kubernetes
-* Basic support for chaincode invoke command
-  
+* Basic support for `chaincode invoke` command
+* Support for `chaincodes list` command (tls and non-tls)
+
 ### Fixes
 * Remove unsupported test library and dependencies
 * Updated Node.js version from 12 to 16 in chaincode